(PC-17812) favorites: put tests next to pages and components
