+++ src/features/favorites/components/FavoritesResults.web.test.tsx
@@ -13,9 +13,9 @@ import {
 import { env } from 'libs/environment'
 import { EmptyResponse } from 'libs/fetch'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
-import { render } from 'tests/utils'
+import { render } from 'tests/utils/web'
 
-import { FavoritesResults } from '../FavoritesResults'
+import { FavoritesResults } from './FavoritesResults'
 
 const mockFavoritesState = initialFavoritesState
 jest.mock('features/favorites/pages/FavoritesWrapper', () => ({
