+++ src/features/favorites/pages/FavoritesSorts.native.test.tsx
@@ -1,6 +1,8 @@
 import React from 'react'
 import waitForExpect from 'wait-for-expect'
 
+import { FavoriteSortBy, FavoritesSorts } from 'features/favorites/pages/FavoritesSorts'
+import { FavoritesWrapper } from 'features/favorites/pages/FavoritesWrapper'
 import { mockGoBack } from 'features/navigation/__mocks__/useGoBack'
 import { analytics } from 'libs/firebase/analytics'
 import {
@@ -10,12 +12,9 @@ import {
   GeoCoordinates,
   GEOLOCATION_USER_ERROR_MESSAGE,
 } from 'libs/geolocation'
-import { superFlushWithAct, fireEvent, render } from 'tests/utils/web'
+import { superFlushWithAct, fireEvent, render } from 'tests/utils'
 
-import { FavoriteSortBy, FavoritesSorts } from '../FavoritesSorts'
-import { FavoritesWrapper } from '../FavoritesWrapper'
-
-jest.mock('../FavoritesWrapper', () => jest.requireActual('../FavoritesWrapper'))
+jest.mock('./FavoritesWrapper', () => jest.requireActual('./FavoritesWrapper'))
 
 const DEFAULT_POSITION = { latitude: 66, longitude: 66 } as GeoCoordinates
 let mockPermissionState = GeolocPermissionState.GRANTED
@@ -23,7 +22,6 @@ let mockPosition: GeoCoordinates | null = DEFAULT_POSITION
 let mockPositionError: GeolocationError | null = null
 const mockTriggerPositionUpdate = jest.fn()
 const mockShowGeolocPermissionModal = jest.fn()
-const mockRequestGeolocPermission = jest.fn()
 
 jest.mock('libs/geolocation/GeolocationWrapper', () => ({
   useGeolocation: () => ({
@@ -32,7 +30,6 @@ jest.mock('libs/geolocation/GeolocationWrapper', () => ({
     positionError: mockPositionError,
     triggerPositionUpdate: mockTriggerPositionUpdate,
     showGeolocPermissionModal: mockShowGeolocPermissionModal,
-    requestGeolocPermission: mockRequestGeolocPermission,
   }),
 }))
 
@@ -52,7 +49,7 @@ describe('FavoritesSorts component', () => {
   it('should go back on validate', async () => {
     const renderAPI = await renderFavoritesSort()
 
-    fireEvent.click(renderAPI.getByText('Valider'))
+    fireEvent.press(renderAPI.getByText('Valider'))
 
     await waitForExpect(() => {
       expect(mockGoBack).toBeCalled()
@@ -74,8 +71,8 @@ describe('FavoritesSorts component', () => {
     }) => {
       const renderAPI = await renderFavoritesSort()
 
-      fireEvent.click(renderAPI.getByText(sortByWording))
-      fireEvent.click(renderAPI.getByText('Valider'))
+      fireEvent.press(renderAPI.getByText(sortByWording))
+      fireEvent.press(renderAPI.getByText('Valider'))
 
       await waitForExpect(() => {
         expect(analytics.logHasAppliedFavoritesSorting).toBeCalledWith({
@@ -93,7 +90,7 @@ describe('FavoritesSorts component', () => {
     }
     const renderAPI = await renderFavoritesSort()
 
-    fireEvent.click(renderAPI.getByText('Proximité géographique'))
+    fireEvent.press(renderAPI.getByText('Proximité géographique'))
 
     renderAPI.getByText(mockPositionError.message)
   })
@@ -101,9 +98,9 @@ describe('FavoritesSorts component', () => {
   it('should trigger analytics=AROUND_ME when clicking on "Proximité géographique" then accepting geoloc then validating', async () => {
     const renderAPI = await renderFavoritesSort()
 
-    fireEvent.click(renderAPI.getByText('Proximité géographique'))
+    fireEvent.press(renderAPI.getByText('Proximité géographique'))
     await superFlushWithAct()
-    fireEvent.click(renderAPI.getByText('Valider'))
+    fireEvent.press(renderAPI.getByText('Valider'))
 
     await waitForExpect(() => {
       expect(
@@ -122,9 +119,9 @@ describe('FavoritesSorts component', () => {
     mockPermissionState = GeolocPermissionState.DENIED
     const renderAPI = await renderFavoritesSort()
 
-    fireEvent.click(renderAPI.getByText('Proximité géographique'))
+    fireEvent.press(renderAPI.getByText('Proximité géographique'))
     await superFlushWithAct()
-    fireEvent.click(renderAPI.getByText('Valider'))
+    fireEvent.press(renderAPI.getByText('Valider'))
 
     await waitForExpect(() => {
       expect(analytics.logHasAppliedFavoritesSorting).toBeCalledWith({
