+++ src/features/favorites/pages/utils/sorts.test.ts
@@ -2,7 +2,7 @@ import { FavoriteResponse } from 'api/gen'
 import { paginatedFavoritesResponseSnap } from 'features/favorites/fixtures/favoritesResponse'
 import { GeoCoordinates } from 'libs/geolocation'
 
-import { sortByIdDesc, sortByAscendingPrice, sortByDistanceAroundMe } from '../sorts'
+import { sortByIdDesc, sortByAscendingPrice, sortByDistanceAroundMe } from './sorts'
 
 const baseOffer: FavoriteResponse = paginatedFavoritesResponseSnap.favorites[0]
